@@ -27,22 +27,22 @@
     "v8Flags": "--expose_gc"
   },
   "dependencies": {
-    "@angular/animations": "~5.2.0",
-    "@angular/common": "~5.2.0",
-    "@angular/compiler": "~5.2.0",
-    "@angular/core": "~5.2.0",
-    "@angular/forms": "~5.2.0",
-    "@angular/http": "~5.2.0",
-    "@angular/platform-browser": "~5.2.0",
-    "@angular/platform-browser-dynamic": "~5.2.0",
-    "@angular/router": "~5.2.0",
-    "nativescript-angular": "~5.2.0",
+    "@angular/animations": "~6.0.0-rc.0",
+    "@angular/common": "~6.0.0-rc.0",
+    "@angular/compiler": "~6.0.0-rc.0",
+    "@angular/core": "~6.0.0-rc.0",
+    "@angular/forms": "~6.0.0-rc.0",
+    "@angular/http": "~6.0.0-rc.0",
+    "@angular/platform-browser": "~6.0.0-rc.0",
+    "@angular/platform-browser-dynamic": "~6.0.0-rc.0",
+    "@angular/router": "~6.0.0-rc.0",
+    "nativescript-angular": "github:nativescript/nativescript-angular#sis0k0/ng6",
     "nativescript-imagepicker": "~5.0.0",
     "nativescript-plugin-firebase": "~5.0.5",
     "nativescript-theme-core": "~1.0.4",
     "nativescript-ui-listview": "~3.5.0",
     "reflect-metadata": "~0.1.10",
-    "rxjs": "~5.5.5",
+    "rxjs": "~6.0.0-rc.0",
     "tns-core-modules": "next",
     "zone.js": "~0.8.18"
   },
@@ -50,7 +50,7 @@
     "codelyzer": "~4.0.2",
     "nativescript-dev-sass": "~1.3.5",
     "nativescript-dev-typescript": "~0.6.0",
-    "nativescript-dev-webpack": "~0.9.1",
+    "nativescript-dev-webpack": "github:nativescript/nativescript-dev-webpack#sis0k0/webpack4",
     "tslint": "~5.8.0",
     "typescript": "~2.7.2"
   },
