@@ -1,5 +1,6 @@
 import { Component, OnInit } from "@angular/core";
 import { PageRoute, RouterExtensions } from "nativescript-angular/router";
+import { switchMap } from "rxjs/operators";
 import { alert } from "ui/dialogs";
 
 import { CarEditService } from "../shared/car-edit.service";
@@ -46,7 +47,7 @@ export class CarDetailEditComponent implements OnInit {
         * http://docs.nativescript.org/angular/core-concepts/angular-navigation.html#passing-parameter
         *************************************************************/
         this._pageRoute.activatedRoute
-            .switchMap((activatedRoute) => activatedRoute.params)
+            .pipe(switchMap((activatedRoute) => activatedRoute.params))
             .forEach((params) => {
                 const carId = params.id;
 
