@@ -1,5 +1,6 @@
 import { Component, OnInit } from "@angular/core";
 import { PageRoute, RouterExtensions } from "nativescript-angular/router";
+import { switchMap } from "rxjs/operators";
 
 import { Car } from "../shared/car.model";
 import { CarService } from "../shared/car.service";
@@ -34,7 +35,7 @@ export class CarDetailComponent implements OnInit {
         * http://docs.nativescript.org/angular/core-concepts/angular-navigation.html#passing-parameter
         *************************************************************/
         this._pageRoute.activatedRoute
-            .switchMap((activatedRoute) => activatedRoute.params)
+            .pipe(switchMap((activatedRoute) => activatedRoute.params))
             .forEach((params) => {
                 const carId = params.id;
 
