@@ -2,6 +2,7 @@ import { Component, OnDestroy, OnInit } from "@angular/core";
 import { ObservableArray } from "data/observable-array";
 import { RouterExtensions } from "nativescript-angular/router";
 import { ListViewEventData } from "nativescript-ui-listview";
+import { finalize } from "rxjs/operators";
 import { Subscription } from "rxjs/Subscription";
 
 import { Car } from "./shared/car.model";
@@ -42,7 +43,7 @@ export class CarListComponent implements OnInit, OnDestroy {
         *************************************************************/
         if (!this._dataSubscription) {
             this._dataSubscription = this._carService.load()
-                .finally(() => this._isLoading = false)
+                .pipe(finalize(() => this._isLoading = false))
                 .subscribe((cars: Array<Car>) => {
                     this._cars = new ObservableArray(cars);
                     this._isLoading = false;
